setup environment vanilla-ts vite
